Se modificó la función de eliminar producto, ahora lo hace por código y no por ID de firebase

+++ assets/js/delete.js
@@ -1,12 +1,21 @@
-import { doc, deleteDoc } from 'https://www.gstatic.com/firebasejs/10.5.2/firebase-firestore.js';
+import { doc, deleteDoc, query, where, getDocs, collection } from 'https://www.gstatic.com/firebasejs/10.5.2/firebase-firestore.js';
 
 import { db } from '../app/firebase.js';
 
-// Función para eliminar información en Firestore
-export const eliminarProducto = async (id) => {
+export const eliminarProducto = async (codigo) => {
   try {
-    const productoDoc = doc(db, 'productos', id);
+    // Buscar el documento con el código proporcionado
+    const q = query(collection(db, 'productos'), where('codigo', '==', codigo));
+    const querySnapshot = await getDocs(q);
 
+    // Verificar si se encontró algún documento con el código
+    if (querySnapshot.size === 0) {
+      console.error('No se encontró ningún producto con el código proporcionado');
+      return;
+    }
+
+    // Eliminar el primer documento encontrado (asumiendo que los códigos son únicos)
+    const productoDoc = querySnapshot.docs[0].ref;
     await deleteDoc(productoDoc);
 
     console.log('Documento eliminado correctamente');
@@ -14,6 +23,7 @@ export const eliminarProducto = async (id) => {
     console.error('Error al eliminar el documento:', error);
   }
 };
+
 
 export const eliminarPopular = async (id) => {
   try {
